[PacketCmd->GamePacketID] PacketReader
* PacketReader:
  * Changed all usages of PacketCmd with GamePacketID & LoadScreenPacketID.
* PacketHeader:
  * Cmd variable changed to GamePacketID.
* PacketHandlerManager:
  * Separated convertorTable variable into game and loadScreen to account for PacketCmd replaced with GamePacketID.
* PacketType:
  * Replaced PacketCmd with GamePacketID.
  * Added LoadScreenPacketID as an option.

+++ GameServerLib/Chatbox/Commands/DebugModeCommand.cs
@@ -40,7 +40,7 @@ namespace LeagueSandbox.GameServer.Chatbox.Commands
         public override void Execute(int userId, bool hasReceivedArguments, string arguments = "")
         {
             _userId = userId;
-            _userChampion = _playerManager.GetPeerInfo((ulong)userId).Champion;
+            _userChampion = _playerManager.GetPeerInfo(userId).Champion;
             if (_debugMode != 0)
             {
                 if (_debugMode == 1)
