+++ GameServerLib/Chatbox/Commands/SkillpointsCommand.cs
@@ -19,7 +19,7 @@ namespace LeagueSandbox.GameServer.Chatbox.Commands
 
         public override void Execute(int userId, bool hasReceivedArguments, string arguments = "")
         {
-            var champion = _playerManager.GetPeerInfo((ulong)userId).Champion;
+            var champion = _playerManager.GetPeerInfo(userId).Champion;
             champion.SkillPoints = 17;
 
             _game.PacketNotifier.NotifyNPC_UpgradeSpellAns(userId, champion.NetId, 0, 0, 17);
