+++ GameServerLib/Packets/PacketHandlers/HandleSurrender.cs
@@ -18,7 +18,7 @@ namespace LeagueSandbox.GameServer.Packets.PacketHandlers
 
         public override bool HandlePacket(int userId, SurrenderRequest req)
         {
-            var c = _pm.GetPeerInfo((ulong)userId).Champion;
+            var c = _pm.GetPeerInfo(userId).Champion;
             _game.Map.MapProperties.HandleSurrender(userId, c, req.VotedYes);
             return true;
         }
